+++ optimism/client.go
@@ -25,7 +25,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/coinbase/rosetta-ethereum/optimism/utilities/artifacts"
 	RosettaTypes "github.com/coinbase/rosetta-sdk-go/types"
 	ethereum "github.com/ethereum-optimism/optimism/l2geth"
 	"github.com/ethereum-optimism/optimism/l2geth/common"
@@ -35,6 +34,7 @@ import (
 	"github.com/ethereum-optimism/optimism/l2geth/params"
 	"github.com/ethereum-optimism/optimism/l2geth/rlp"
 	"github.com/ethereum-optimism/optimism/l2geth/rpc"
+	"github.com/inphi/optimism-rosetta/optimism/utilities/artifacts"
 
 	"github.com/ethereum/go-ethereum/crypto"
 	"golang.org/x/sync/semaphore"
@@ -278,31 +278,6 @@ func toBlockNumArg(number *big.Int) string {
 	return hexutil.EncodeBig(number)
 }
 
-// Block returns a populated block at the *RosettaTypes.PartialBlockIdentifier.
-// If neither the hash or index is populated in the *RosettaTypes.PartialBlockIdentifier,
-// the current block is returned.
-func (ec *Client) Block(
-	ctx context.Context,
-	blockIdentifier *RosettaTypes.PartialBlockIdentifier,
-) (*RosettaTypes.Block, error) {
-	if blockIdentifier != nil {
-		if blockIdentifier.Hash != nil {
-			return ec.getParsedBlock(ctx, "eth_getBlockByHash", *blockIdentifier.Hash, true)
-		}
-
-		if blockIdentifier.Index != nil {
-			return ec.getParsedBlock(
-				ctx,
-				"eth_getBlockByNumber",
-				toBlockNumArg(big.NewInt(*blockIdentifier.Index)),
-				true,
-			)
-		}
-	}
-
-	return ec.getParsedBlock(ctx, "eth_getBlockByNumber", toBlockNumArg(nil), true)
-}
-
 // Header returns a block header from the current canonical chain. If number is
 // nil, the latest known header is returned.
 func (ec *Client) blockHeader(ctx context.Context, number *big.Int) (*types.Header, error) {
@@ -313,103 +288,6 @@ func (ec *Client) blockHeader(ctx context.Context, number *big.Int) (*types.Head
 	}
 
 	return head, err
-}
-
-type rpcBlock struct {
-	Hash         common.Hash      `json:"hash"`
-	Transactions []rpcTransaction `json:"transactions"`
-	UncleHashes  []common.Hash    `json:"uncles"`
-}
-
-func (ec *Client) getBlock(
-	ctx context.Context,
-	blockMethod string,
-	args ...interface{},
-) (
-	*types.Block,
-	[]*loadedTransaction,
-	error,
-) {
-	var raw json.RawMessage
-	err := ec.c.CallContext(ctx, &raw, blockMethod, args...)
-	if err != nil {
-		return nil, nil, fmt.Errorf("%w: block fetch failed", err)
-	} else if len(raw) == 0 {
-		return nil, nil, ethereum.NotFound
-	}
-
-	// Decode header and transactions
-	var head types.Header
-	var body rpcBlock
-	if err := json.Unmarshal(raw, &head); err != nil {
-		return nil, nil, err
-	}
-	if err := json.Unmarshal(raw, &body); err != nil {
-		return nil, nil, err
-	}
-
-	// Get all transaction receipts
-	receipts, err := ec.getBlockReceipts(ctx, body.Hash, body.Transactions)
-	if err != nil {
-		return nil, nil, fmt.Errorf("%w: could not get receipts for %x", err, body.Hash[:])
-	}
-
-	// Get block traces (not possible to make idempotent block transaction trace requests)
-	//
-	// We fetch traces last because we want to avoid limiting the number of other
-	// block-related data fetches we perform concurrently (we limit the number of
-	// concurrent traces that are computed to 16 to avoid overwhelming geth).
-	var traces []*Call
-	var addTraces bool
-	if head.Number.Int64() != GenesisBlockIndex { // not possible to get traces at genesis
-		addTraces = true
-		traces, err = ec.getTransactionTraces(ctx, body.Transactions)
-		if err != nil {
-			return nil, nil, fmt.Errorf("%w: could not get traces for all txs in block %x", err, body.Hash[:])
-		}
-	}
-
-	// Convert all txs to loaded txs
-	txs := make([]*types.Transaction, len(body.Transactions))
-	loadedTxs := make([]*loadedTransaction, len(body.Transactions))
-	for i, tx := range body.Transactions {
-		txs[i] = tx.tx
-		receipt := receipts[i]
-
-		var feeAmount *big.Int
-		if feeAmountInDupTx := originalFeeAmountInDupTx[string(body.Hash.Hex())]; feeAmountInDupTx == "" {
-			gasUsedBig := new(big.Int).SetUint64(receipt.GasUsed)
-			l2feeAmount := gasUsedBig.Mul(gasUsedBig, txs[i].GasPrice())
-			feeAmount = l2feeAmount.Add(l2feeAmount, receipts[i].L1Fee)
-		} else {
-			// The fees reported in the tx receipt refers to the succeeding duplicate tx rather thaan the original.
-			// We fix the feeAmount here to use the original so that balances are accounted for
-			// Note that these duplicate transactions all failed to complete, so there aren't any additional mint/burn operations to account for.
-			feeAmount = hexutil.MustDecodeBig(feeAmountInDupTx)
-		}
-
-		loadedTxs[i] = tx.LoadedTransaction()
-		loadedTxs[i].Transaction = txs[i]
-		loadedTxs[i].FeeAmount = feeAmount
-		// Miner is fixed on Optimism and block rewards are sent internally to the OVM_SEQUENCER_FEE_VAULT contract.
-		// However, the block.coinbase is set to 0x0, rather than the vault contract.
-		// It would be nice for l2geth to populate the appropriate coinbase so we're robust against changes to the vault addresss.
-		loadedTxs[i].Miner = sequencerFeeVaultAddr
-		loadedTxs[i].Receipt = receipt
-		loadedTxs[i].Status = receipt.Status == 1
-
-		// Continue if calls does not exist (occurs at genesis)
-		if !addTraces {
-			continue
-		}
-
-		loadedTxs[i].Trace = traces[i]
-	}
-
-	return types.NewBlockWithHeader(&head).WithBody(
-		txs,
-		nil, // Sequencer blocks do not have uncles with instant confirmation
-	), loadedTxs, nil
 }
 
 func (ec *Client) getTransactionTraces(
@@ -503,7 +381,7 @@ func (ec *Client) getBlockReceipts(
 func (ec *Client) erc20TokenOps(
 	ctx context.Context,
 	block *types.Block,
-	tx *loadedTransaction,
+	tx *legacyTransaction,
 	startIndex int,
 ) ([]*RosettaTypes.Operation, error) {
 	receipt := tx.Receipt
@@ -709,7 +587,7 @@ func containsTopic(log *types.Log, topic string) bool {
 
 // traceOps returns all *RosettaTypes.Operation for a given
 // array of flattened traces.
-func traceOps(block *types.Block, calls []*flatCall, startIndex int) []*RosettaTypes.Operation { // nolint: gocognit
+func traceOps(block *types.Block, calls []*FlatCall, startIndex int) []*RosettaTypes.Operation { // nolint: gocognit
 	var ops []*RosettaTypes.Operation
 	if len(calls) == 0 {
 		return ops
@@ -1153,141 +1031,8 @@ func validateCallInput(params map[string]interface{}) (*GetCallInput, error) {
 	return &input, nil
 }
 
-func (ec *Client) getParsedBlock(
-	ctx context.Context,
-	blockMethod string,
-	args ...interface{},
-) (
-	*RosettaTypes.Block,
-	error,
-) {
-	// TODO: dispatch based on if it's pre or post bedrock
-	block, loadedTransactions, err := ec.getBlock(ctx, blockMethod, args...)
-	if err != nil {
-		return nil, fmt.Errorf("%w: could not get block", err)
-	}
-
-	blockIdentifier := &RosettaTypes.BlockIdentifier{
-		Hash:  block.Hash().String(),
-		Index: block.Number().Int64(),
-	}
-
-	parentBlockIdentifier := blockIdentifier
-	if blockIdentifier.Index != GenesisBlockIndex {
-		parentBlockIdentifier = &RosettaTypes.BlockIdentifier{
-			Hash:  block.ParentHash().Hex(),
-			Index: blockIdentifier.Index - 1,
-		}
-	}
-
-	txs, err := ec.populateTransactions(ctx, blockIdentifier, block, loadedTransactions)
-	if err != nil {
-		return nil, err
-	}
-
-	return &RosettaTypes.Block{
-		BlockIdentifier:       blockIdentifier,
-		ParentBlockIdentifier: parentBlockIdentifier,
-		Timestamp:             convertTime(block.Time()),
-		Transactions:          txs,
-	}, nil
-}
-
 func convertTime(time uint64) int64 {
 	return int64(time) * 1000
-}
-
-func (ec *Client) populateTransactions(
-	ctx context.Context,
-	blockIdentifier *RosettaTypes.BlockIdentifier,
-	block *types.Block,
-	loadedTransactions []*loadedTransaction,
-) ([]*RosettaTypes.Transaction, error) {
-	transactions := make(
-		[]*RosettaTypes.Transaction,
-		len(block.Transactions()),
-	)
-
-	for i, tx := range loadedTransactions {
-		if tx.From != nil && tx.Transaction != nil && tx.Transaction.To() != nil {
-			from, to := tx.From.Hex(), tx.Transaction.To().Hex()
-
-			// These are tx across L1 and L2. These cost zero gas as they're manufactured by the sequencer
-			if from == zeroAddr {
-				tx.FeeAmount.SetUint64(0)
-			} else if (to == gasPriceOracleAddr.Hex()) && (from == gasPriceOracleOwnerMainnet.Hex() || from == gasPriceOracleOwnerKovan.Hex() || from == gasPriceOracleOwnerGoerli.Hex()) {
-				// The sequencer doesn't charge the owner of the gpo.
-				// Set the fee mount to zero to not affect gpo owner balances
-				tx.FeeAmount.SetUint64(0)
-			}
-		}
-
-		transaction, err := ec.populateTransaction(ctx, block, tx)
-		if err != nil {
-			return nil, fmt.Errorf("%w: cannot parse %s", err, tx.Transaction.Hash().Hex())
-		}
-
-		transactions[i] = transaction
-	}
-
-	return transactions, nil
-}
-
-func (ec *Client) populateTransaction(
-	ctx context.Context,
-	block *types.Block,
-	tx *loadedTransaction,
-) (*RosettaTypes.Transaction, error) {
-	ops := []*RosettaTypes.Operation{}
-
-	// Compute fee operations
-	feeOps := feeOps(tx)
-	patchFeeOps(ec.p.ChainID, block, tx.Transaction, feeOps)
-	ops = append(ops, feeOps...)
-
-	erc20TokenOps, err := ec.erc20TokenOps(ctx, block, tx, len(ops))
-	if err != nil {
-		return nil, err
-	}
-	ops = append(ops, erc20TokenOps...)
-
-	traces := flattenTraces(tx.Trace, []*flatCall{})
-
-	traceOps := traceOps(block, traces, len(ops))
-	ops = append(ops, traceOps...)
-
-	// Marshal receipt and trace data
-	// TODO: replace with marshalJSONMap (used in `services`)
-	receiptBytes, err := tx.Receipt.MarshalJSON()
-	if err != nil {
-		return nil, fmt.Errorf("%w: cannot marshal receipt json", err)
-	}
-
-	var receiptMap map[string]interface{}
-	if err := json.Unmarshal(receiptBytes, &receiptMap); err != nil {
-		return nil, fmt.Errorf("%w: cannot unmarshal receipt bytes into map", err)
-	}
-
-	// TODO: Currently not saving raw trace
-	// var traceMap map[string]interface{}
-	// if err := json.Unmarshal(tx.RawTrace, &traceMap); err != nil {
-	// 	return nil, fmt.Errorf("%w: cannot unmarshal raw trace", err)
-	// }
-
-	populatedTransaction := &RosettaTypes.Transaction{
-		TransactionIdentifier: &RosettaTypes.TransactionIdentifier{
-			Hash: tx.Transaction.Hash().Hex(),
-		},
-		Operations: ops,
-		Metadata: map[string]interface{}{
-			"gas_limit": hexutil.EncodeUint64(tx.Transaction.Gas()),
-			"gas_price": hexutil.EncodeBig(tx.Transaction.GasPrice()),
-			"receipt":   receiptMap,
-			// "trace":     traceMap, // TODO: use non-raw trace
-		},
-	}
-
-	return populatedTransaction, nil
 }
 
 type rpcProgress struct {
