--- optimism/types/doc.go
@@ -1,2 +1,0 @@
-// Package types provides types for the optimism client.
-package types
