+++ services/construction_service_test.go
@@ -23,9 +23,9 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/coinbase/rosetta-ethereum/configuration"
-	mocks "github.com/coinbase/rosetta-ethereum/mocks/services"
-	"github.com/coinbase/rosetta-ethereum/optimism"
+	"github.com/inphi/optimism-rosetta/configuration"
+	mocks "github.com/inphi/optimism-rosetta/mocks/services"
+	"github.com/inphi/optimism-rosetta/optimism"
 
 	"github.com/coinbase/rosetta-sdk-go/types"
 	ethereum "github.com/ethereum-optimism/optimism/l2geth"
